put all icesat-2 specific functions into a reference module, is2ref
